pyproject.toml
[project]
name = "personal-assistant"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "slack-bolt>=1.18.0",
    "anthropic>=0.25.0",
    "python-dotenv>=1.0.0",
    "apscheduler>=3.10.0",
    "pyobjc-framework-EventKit>=10.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
